[ADD] Implemented map on a page
