--- src/components/MapComponent/MapComponent.hbs
@@ -1,9 +1,0 @@
-<div class="map">
-{{!-- {{{Map}}} --}}
-<div class="whatever"> Sample text</div>
-    <div id="mapid2" style="width: 100%; height: 100%;"  tabindex="1"> 
-    </div>
-
-{{!-- <div id="mapid" style="width: 100%; height: 100%;"  tabindex="1"> 
-    </div> --}}
-</div>
