--- src/components/MapComponent/_MapComponent.scss
@@ -1,12 +1,0 @@
-.gps-coords-for-root__input {
-  font-size: 1rem;
-  display: flex;
-  flex-direction: column;
-  width: 30em;
-}
-
-.gps-coords-for-root-form__admission {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-}
