+++ src/components/UseMapForm/UseMapForm.hbs
@@ -1,17 +1,5 @@
-<div class="map-usage__form-block" id='custom_div1'>
-    {{!-- <form class="gps-coords-for-root__input">
-        {{#each input }}
-        {{{ this }}}
-        {{/each}}
-    </form>
-    <div class="gps-coords-for-root-form__admission">
-        {{{ StandardButton }}}
-    </div> --}}
-    <div id="mapid">
-        {{{MapForm}}}   
-    </div>
-    
-    <div id="mapid" style="width: 100%; height: 100%;" class="leaflet-container" tabindex="1">
-        {{{ Map }}} 
+<div class="map-usage__form-block">
+    <link rel="stylesheet" href="https://unpkg.com/leaflet@1.7.1/dist/leaflet.css" integrity="sha512-xodZBNTC5n17Xt2atTPuE1HxjVMSvLVW9ocqUKLsCC5CXdbqCmblAshOMAS6/keqq/sMZMZ19scR4PsZChSR7A==" crossorigin="">
+    <div class="map-container" id="mapid" style="width: 700px; height: 500px;" tabindex="1">
     </div>
 </div>
