+++ src/components/UseMapForm/UseMapForm.js
@@ -15,13 +15,29 @@ export default class UseMapForm extends BaseComponent {
 
         this._context.RegisterPath = Routes.UseMap;
         this._context.RegisterEvent = Events.ChangePath;
+    }
 
-        let elementMapId = document.createElement("div");
+draw_map() {
+        let element = document.getElementById("mapid");
 
-        elementMapId.id = "mapid";
-        elementMapId.style = "width: 100%; height: 100%;";
-        elementMapId.tabindex = "1";
-        this._context.MapForm = elementMapId.context;
-        // document.append(elementMapId);
-        }
+        console.log("found maps");
+        element = L.map('mapid').setView([51.505, -0.09], 13);
+        L.tileLayer('https://api.mapbox.com/styles/v1/{id}/tiles/{z}/{x}/{y}?access_token=pk.eyJ1IjoibWFwYm94IiwiYSI6ImNpejY4NXVycTA2emYycXBndHRqcmZ3N3gifQ.rJcFIG214AriISLbB6B5aw', {
+            maxZoom: 18,
+            attribution: 'Map data &copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors, ' +
+            'Imagery © <a href="https://www.mapbox.com/">Mapbox</a>',
+            id: 'mapbox/streets-v11',
+            tileSize: 512,
+            zoomOffset: -1
+        }).addTo(element);
+
+        // function onMapClick(e) {
+        //     popup
+        //         .setLatLng(e.latlng)
+        //         .setContent("You clicked the map at " + e.latlng.toString())
+        //         .openOn(mymap);
+        // }
+
+        // mymap.on('click', onMapClick);
+    }
 }
